Move types to another file
